TVGUI-17 Move styling for pages into sub-directory
