+++ src/Pages/ShowListPage.tsx
@@ -5,7 +5,7 @@ import { DeleteOutlined } from "@ant-design/icons";
 import AddShow from "../components/AddShow";
 import { SearchListContext, UserContext } from "../contexts";
 import { removeShowFromList } from "../requests/requests";
-import '../ShowList.css';
+import './styles/ShowList.css';
 
 const ShowListPage = () => {
     const [addingNewShow, setAddingNewShow] = useState(false);
